requirements.txt
# Core dependencies
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.3.0

# NLP and ML
datasets>=2.15.0
accelerate>=0.25.0
sentencepiece>=0.1.99
sacremoses>=0.0.53
transformers>=4.36.0
torch>=2.1.0

# API and serving
fastapi>=0.105.0
uvicorn>=0.24.0
pydantic>=2.5.0

# Data collection
praw>=7.7.0
python-dotenv>=1.0.0

# Monitoring and MLOps
mlflow>=2.9.0
prometheus-client>=0.19.0
psutil>=5.9.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Utilities
tqdm>=4.66.0
emoji>=2.8.0

# Development Dependencies
black==23.12.1
flake8==7.0.0
pytest>=8.3.0
pytest-cov>=6.0.0
httpx
pytest-mock